# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CharacterDetail": {
        "additionalProperties": true,
        "description": "对应 api.characters._serialize_character 返回。和 CharacterBrief 的区别：\nBrief 只给 DM 上下文用（精简），Detail 给前端创角 / 角色面板（完整）。",
        "properties": {
          "ability_scores": {
            "additionalProperties": true,
            "default": {},
            "title": "Ability Scores",
            "type": "object"
          },
          "ac": {
            "title": "Ac",
            "type": "integer"
          },
          "alignment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment"
          },
          "background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background"
          },
          "backstory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backstory"
          },
          "cantrips": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cantrips",
            "type": "array"
          },
          "cantrips_count": {
            "default": 0,
            "title": "Cantrips Count",
            "type": "integer"
          },
          "caster_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caster Type"
          },
          "catchphrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catchphrase"
          },
          "char_class": {
            "title": "Char Class",
            "type": "string"
          },
          "combat_preference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combat Preference"
          },
          "concentration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concentration"
          },
          "condition_durations": {
            "additionalProperties": true,
            "default": {},
            "title": "Condition Durations",
            "type": "object"
          },
          "conditions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "death_saves": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Death Saves"
          },
          "derived": {
            "additionalProperties": true,
            "default": {},
            "title": "Derived",
            "type": "object"
          },
          "equipment": {
            "additionalProperties": true,
            "default": {},
            "title": "Equipment",
            "type": "object"
          },
          "feats": {
            "default": [],
            "items": {},
            "title": "Feats",
            "type": "array"
          },
          "fighting_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fighting Style"
          },
          "hp_current": {
            "title": "Hp Current",
            "type": "integer"
          },
          "hp_max": {
            "title": "Hp Max",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_player": {
            "title": "Is Player",
            "type": "boolean"
          },
          "known_spells": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Known Spells",
            "type": "array"
          },
          "languages": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Languages",
            "type": "array"
          },
          "level": {
            "title": "Level",
            "type": "integer"
          },
          "multiclass_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multiclass Info"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "personality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personality"
          },
          "prepared_spells": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Prepared Spells",
            "type": "array"
          },
          "proficient_saves": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Proficient Saves",
            "type": "array"
          },
          "proficient_skills": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Proficient Skills",
            "type": "array"
          },
          "race": {
            "title": "Race",
            "type": "string"
          },
          "speech_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speech Style"
          },
          "spell_slots": {
            "additionalProperties": true,
            "default": {},
            "title": "Spell Slots",
            "type": "object"
          },
          "spell_slots_max": {
            "additionalProperties": true,
            "default": {},
            "title": "Spell Slots Max",
            "type": "object"
          },
          "subclass": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subclass"
          },
          "subclass_effects": {
            "additionalProperties": true,
            "default": {},
            "title": "Subclass Effects",
            "type": "object"
          },
          "tool_proficiencies": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tool Proficiencies",
            "type": "array"
          }
        },
        "required": [
          "id",
          "is_player",
          "name",
          "race",
          "char_class",
          "level",
          "hp_current",
          "hp_max",
          "ac"
        ],
        "title": "CharacterDetail",
        "type": "object"
      },
...
      "ConditionRequest": {
        "properties": {
          "condition": {
            "title": "Condition",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "is_enemy": {
            "default": false,
            "title": "Is Enemy",
            "type": "boolean"
          },
          "rounds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds"
          }
        },
        "required": [
          "entity_id",
          "condition"
        ],
        "title": "ConditionRequest",
        "type": "object"
      },
...
      "CreateCharacterRequest": {
        "properties": {
          "ability_scores": {
            "$ref": "#/components/schemas/AbilityScores"
          },
          "alignment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alignment"
          },
          "background": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background"
          },
          "backstory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backstory"
          },
          "bonus_languages": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Bonus Languages",
            "type": "array"
          },
          "cantrips": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cantrips",
            "type": "array"
          },
          "catchphrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catchphrase"
          },
          "char_class": {
            "title": "Char Class",
            "type": "string"
          },
          "combat_preference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combat Preference"
          },
          "equipment_choice": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment Choice"
          },
          "feats": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Feats",
            "type": "array"
          },
          "fighting_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fighting Style"
          },
          "known_spells": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Known Spells",
            "type": "array"
          },
          "level": {
            "default": 1,
            "title": "Level",
            "type": "integer"
          },
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "multiclass_info": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multiclass Info"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "personality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personality"
          },
          "proficient_skills": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Proficient Skills",
            "type": "array"
          },
          "race": {
            "title": "Race",
            "type": "string"
          },
          "speech_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speech Style"
          },
          "subclass": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subclass"
          }
        },
        "required": [
          "module_id",
          "name",
          "race",
          "char_class",
          "ability_scores"
        ],
        "title": "CreateCharacterRequest",
        "type": "object"
      },
...
      "ExhaustionRequest": {
        "properties": {
          "change": {
            "default": 1,
            "title": "Change",
            "type": "integer"
          }
        },
        "title": "ExhaustionRequest",
        "type": "object"
      },
      "ExhaustionUpdateResult": {
        "additionalProperties": true,
        "description": "PATCH /characters/{id}/exhaustion",
        "properties": {
          "effects": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Effects",
            "type": "array"
          },
          "exhaustion_level": {
            "title": "Exhaustion Level",
            "type": "integer"
          },
          "is_dead": {
            "default": false,
            "title": "Is Dead",
            "type": "boolean"
          }
        },
        "required": [
          "exhaustion_level"
        ],
        "title": "ExhaustionUpdateResult",
        "type": "object"
      },
...
      "GeneratePartyRequest": {
        "properties": {
          "module_id": {
            "title": "Module Id",
            "type": "string"
          },
          "party_size": {
            "default": 4,
            "title": "Party Size",
            "type": "integer"
          },
          "player_character_id": {
            "title": "Player Character Id",
            "type": "string"
          }
        },
        "required": [
          "module_id",
          "player_character_id"
        ],
        "title": "GeneratePartyRequest",
        "type": "object"
      },
      "GeneratePartyResponse": {
        "additionalProperties": true,
        "description": "POST /characters/generate-party 返回。",
        "properties": {
          "companions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CharacterDetail"
            },
            "title": "Companions",
            "type": "array"
          }
        },
        "title": "GeneratePartyResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LevelUpRequest": {
        "properties": {
          "ability_score_increases": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ability Score Increases"
          },
          "feat_choice": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feat Choice"
          },
          "use_average_hp": {
            "default": true,
            "title": "Use Average Hp",
            "type": "boolean"
          }
        },
        "title": "LevelUpRequest",
        "type": "object"
      },
      "LevelUpResult": {
        "additionalProperties": true,
        "description": "POST /characters/{id}/level-up",
        "properties": {
          "character": {
            "$ref": "#/components/schemas/CharacterDetail"
          },
          "level_up_details": {
            "additionalProperties": true,
            "default": {},
            "title": "Level Up Details",
            "type": "object"
          }
        },
        "required": [
          "character"
        ],
        "title": "LevelUpResult",
        "type": "object"
      },
...
      "PreparedSpellsRequest": {
        "properties": {
          "prepared_spells": {
            "items": {
              "type": "string"
            },
            "title": "Prepared Spells",
            "type": "array"
          }
        },
        "required": [
          "prepared_spells"
        ],
        "title": "PreparedSpellsRequest",
        "type": "object"
      },
      "PreparedSpellsResult": {
        "additionalProperties": true,
        "description": "PATCH /characters/{id}/prepared-spells",
        "properties": {
          "max_prepared": {
            "title": "Max Prepared",
            "type": "integer"
          },
          "prepared_spells": {
            "items": {
              "type": "string"
            },
            "title": "Prepared Spells",
            "type": "array"
          }
        },
        "required": [
          "prepared_spells",
          "max_prepared"
        ],
        "title": "PreparedSpellsResult",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "DnD 5e AI跑团平台后端",
    "title": "AI TRPG Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/characters/create": {
      "post": {
        "operationId": "create_character_characters_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCharacterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Character",
        "tags": [
          "characters"
        ]
      }
    },
    "/characters/generate-party": {
      "post": {
        "operationId": "generate_party_characters_generate_party_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePartyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratePartyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Party",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/characters/{character_id}": {
      "get": {
        "operationId": "get_character_characters__character_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Character",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/characters/{character_id}/exhaustion": {
      "patch": {
        "operationId": "update_exhaustion_characters__character_id__exhaustion_patch",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExhaustionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExhaustionUpdateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Exhaustion",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/characters/{character_id}/level-up": {
      "post": {
        "operationId": "level_up_characters__character_id__level_up_post",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LevelUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LevelUpResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Level Up",
        "tags": [
          "characters"
        ]
      }
    },
    "/characters/{character_id}/prepared-spells": {
      "patch": {
        "operationId": "update_prepared_spells_characters__character_id__prepared_spells_patch",
        "parameters": [
          {
            "in": "path",
            "name": "character_id",
            "required": true,
            "schema": {
              "title": "Character Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreparedSpellsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparedSpellsResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Prepared Spells",
        "tags": [
          "characters"
        ]
      }
    },
...
    "/game/combat/{session_id}/condition/add": {
      "post": {
        "operationId": "add_condition_game_combat__session_id__condition_add_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConditionUpdateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Condition",
        "tags": [
          "combat"
        ]
      }
    },
    "/game/combat/{session_id}/condition/remove": {
      "post": {
        "operationId": "remove_condition_game_combat__session_id__condition_remove_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConditionUpdateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Condition",
        "tags": [
          "combat"
        ]
      }
    },
...
  }
}